user-service/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.7
PyJWT==2.8.0
pytest==7.4.0